chore(loop): split tick_jobs.py into job_identity/scanner_factories/domain_jobs/global_scanner_factories
Pure relocation of the 1370-LOC tick_jobs god module into four cohesive
modules, no behaviour change. Acyclic layering global_scanner_factories →
domain_jobs → scanner_factories → job_identity (leaf):

- job_identity: _ScannerJob, Domain, PER_OVERLAY_DOMAINS and the canonical
  overlay/tuple constants (the leaf shared by all three).
- scanner_factories: the per-overlay _*_scanner_for constructors plus the
  effective-settings / identity-alias / backend-host helpers.
- domain_jobs: the domain slices, jobs_for_domain, the dispatch dicts and
  the per-job run/notify helpers.
- global_scanner_factories: the global-cadence builders plus
  build_default_jobs / build_default_scanners.

Deviation from the PR5 spec's 2-file target: four files, not two. The
500-LOC module-health ratchet forces it — scanner_factories alone is 488
counted LOC, so a 2-way cut cannot stay under the cap. Net structure is
identical (same public surface, re-exported through tick.py).

mock.patch targets repointed to the module where each symbol is now
looked up (def-site for builders, domain_jobs for the dispatch-exercising
patches). tick.py keeps the load_config/discover_overlays re-export so
the legacy monkeypatch path stays live. docs/blueprint loop-topology and
configuration updated to the new module names.
