+++ docs/blueprint/loop-topology.md
@@ -66,7 +66,7 @@ The statusline file is **display-only**, never an orchestration channel; tick st
 
 **Ticket dispositions (phase 7).** `TicketDispositionScanner` walks active pre-PR `Ticket` rows once per tick — states `not_started..reviewed`, active overlay — calling `CodeHostBackend.get_issue(ticket.issue_url)` to detect drift between local intent and remote reality. Three reasons trigger a `ticket.disposition_candidate` signal: `issue_closed` (remote state ∈ `{closed, completed, cancelled}`, often a different PR's `Fixes #N` keyword auto-closed it), `unassigned` (the active user dropped from assignees), `label_removed` (no configured `ready_labels` remain). Multiple reasons emit multiple signals so the statusline shows each independently. **Detection-only**: it never transitions the ticket or tears down worktrees — disposition is a judgment call (was the work superseded? does the colleague's reassignment override mine?) and is uncommon enough that asking the operator is cheap. Tickets past `REVIEWED` are skipped: once a PR exists, `TicketCompletionScanner` covers post-PR state.
 
-**Reassign noise filter (#975).** One operator may carry several handles across platforms (a GitHub login, a GitLab username, an internal id); a reassignment between two is plumbing, not an actionable handoff. `UserSettings.user_identity_aliases` (default `[]`, in `[teatree]`, per-overlay overridable via `OVERLAY_OVERRIDABLE_SETTINGS`) lists them; the scanner drops the `unassigned` signal when `old_owner ∈ aliases` AND every `new_owners` entry ∈ aliases. Reassigns crossing the alias boundary (alias↔colleague, alias→mixed) still render — the colleague handoff is the actionable case. Default empty list keeps legacy behaviour. **Implicit self-group from `backend.identities` (#1113).** When neither `user_identity_aliases` nor an explicit `identity_alias_groups` resolves, `tick_jobs._jobs_for_backend_hosts` unions `backend.identities` (the #976 list) into the scanner's `identity_alias_groups` as an implicit one-human self-group, closing the gap where aliases are empty but identities known. Explicit group config takes precedence.
+**Reassign noise filter (#975).** One operator may carry several handles across platforms (a GitHub login, a GitLab username, an internal id); a reassignment between two is plumbing, not an actionable handoff. `UserSettings.user_identity_aliases` (default `[]`, in `[teatree]`, per-overlay overridable via `OVERLAY_OVERRIDABLE_SETTINGS`) lists them; the scanner drops the `unassigned` signal when `old_owner ∈ aliases` AND every `new_owners` entry ∈ aliases. Reassigns crossing the alias boundary (alias↔colleague, alias→mixed) still render — the colleague handoff is the actionable case. Default empty list keeps legacy behaviour. **Implicit self-group from `backend.identities` (#1113).** When neither `user_identity_aliases` nor an explicit `identity_alias_groups` resolves, `scanner_factories._jobs_for_backend_hosts` unions `backend.identities` (the #976 list) into the scanner's `identity_alias_groups` as an implicit one-human self-group, closing the gap where aliases are empty but identities known. Explicit group config takes precedence.
 
 **Ticket completion (phase 8).** `TicketCompletionScanner` covers the gap dispositions leave: post-ship tickets (states `shipped`, `in_review`, `merged`) whose upstream issue indicates all work is done. Unlike dispositions (detection-only because pre-PR closure is ambiguous), post-ship completion is mechanical — the user shipped, so a closed issue means the work landed. The scanner calls `OverlayBase.is_issue_done(issue_data)` per overlay: default (GitHub) checks `state ∈ {closed, completed}`; GitLab overlays override to check for a process label (e.g. `Process:DEV Review`), since multi-repo tickets have multiple MRs and the issue stays open until the last merges. The dispatcher routes `ticket.completion_detected` as a `mechanical` action — the tick transitions the ticket `request_review → mark_merged → retrospect` without agent involvement. `t3 <overlay> ticket sync-completions` runs the same logic as a one-off bulk cleanup.
