+++ src/teatree/loop/sweep_on_demand.py
@@ -43,7 +43,7 @@ def trigger_sweep_for_verdict(*, slug: str, pr_id: int, overlay: str) -> MergeAt
 
 def _sweep_scanner_for_overlay(overlay: str) -> PrSweepScanner | None:
     from teatree.core.backend_factory import iter_overlay_backends  # noqa: PLC0415
-    from teatree.loop.tick_jobs import _pr_sweep_scanner_for, _user_slack_id_for_overlay  # noqa: PLC0415
+    from teatree.loop.scanner_factories import _pr_sweep_scanner_for, _user_slack_id_for_overlay  # noqa: PLC0415
 
     backend = next((candidate for candidate in iter_overlay_backends() if candidate.name == overlay), None)
     if backend is None:
