+++ src/teatree/loops/eval_local/loop.py
@@ -5,11 +5,12 @@ from typing import TYPE_CHECKING
 from teatree.loops.base import MiniLoop
 
 if TYPE_CHECKING:
-    from teatree.loop.tick_jobs import _ScannerJob
+    from teatree.loop.job_identity import _ScannerJob
 
 
 def _build_jobs(**_: object) -> "list[_ScannerJob]":
-    from teatree.loop.tick_jobs import _eval_local_scanner, _ScannerJob  # noqa: PLC0415
+    from teatree.loop.global_scanner_factories import _eval_local_scanner  # noqa: PLC0415
+    from teatree.loop.job_identity import _ScannerJob  # noqa: PLC0415
 
     scanner = _eval_local_scanner()
     if scanner is None:
